Fix missing WEB from ClientProviderV4, relocate WebEmbeddedWithThumbnail to where it should be
