+++ v2/src/main/java/dev/lavalink/youtube/clients/WebEmbeddedWithThumbnail.java
@@ -1,8 +1,6 @@
-package dev.lavalink.youtube.clients.skeleton;
+package dev.lavalink.youtube.clients;
 
-import dev.lavalink.youtube.clients.ClientOptions;
-import dev.lavalink.youtube.clients.Web;
-import dev.lavalink.youtube.clients.WebEmbedded;
+import dev.lavalink.youtube.clients.skeleton.NonMusicClientWithThumbnail;
 import org.jetbrains.annotations.NotNull;
 
 public class WebEmbeddedWithThumbnail extends WebEmbedded implements NonMusicClientWithThumbnail {
